fix: just separate data-table card for Transformed data-table
